Small refactoring : renamed the Ragger client
